requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.6
Flask-Login>=0.6.3
boto3==1.28.57
gunicorn==21.2.0
python-dotenv==1.0.0 
eventlet>=0.36.1
